Fixed requirements.txt.
Fixed dbl webhook.py
