+++ cogs/bot.py
@@ -3,12 +3,11 @@ from disnake.enums import ButtonStyle
 from disnake.ext import commands
 from disnake.ui import Button
 
-from main import UndertaleBot
 from utility.constants import BLUE
 
 
 class Bot(commands.Cog):
-    def __init__(self, bot: UndertaleBot):
+    def __init__(self, bot):
         self.bot = bot
 
     @commands.slash_command()
